closes #1 | error on the pin(9)
The Servo library supports up to 12 motors on most Arduino boards and 48 on the Arduino Mega. On boards other than the Mega, use of the library disables analogWrite() (PWM) functionality on pins 9 and 10, whether or not there is a Servo on those pins. On the Mega, up to 12 servos can be used without interfering with PWM functionality; use of 12 to 23 motors will disable PWM on pins 11 and 12.

ref: https://stackoverflow.com/questions/75318922/motor-driver-isnt-working-properly-while-servo-attacha0-line-present-in-the/75319114#75319114